Make BitsTypeDefinition.Bit.position() return Uint32
We seem to be brdiging back and forth between long and Uint32,
as PositionEffectiveStatement is defined in terms of Uint32. Bite
the bullet and convert Bit, too, removing a number of impossible
cases.

JIRA: YANGTOOLS-1072

+++ yang/yang-parser-rfc7950/src/test/java/org/opendaylight/yangtools/yang/parser/stmt/rfc7950/Bug6887Test.java
@@ -18,6 +18,7 @@ import java.util.List;
 import org.junit.Test;
 import org.opendaylight.yangtools.yang.common.QName;
 import org.opendaylight.yangtools.yang.common.Revision;
+import org.opendaylight.yangtools.yang.common.Uint32;
 import org.opendaylight.yangtools.yang.model.api.LeafSchemaNode;
 import org.opendaylight.yangtools.yang.model.api.Module;
 import org.opendaylight.yangtools.yang.model.api.SchemaContext;
@@ -302,7 +303,7 @@ public class Bug6887Test {
     }
 
     private static Bit createBit(final String name, final long position) {
-        return BitBuilder.create(name, position).build();
+        return BitBuilder.create(name, Uint32.valueOf(position)).build();
     }
 
     private static void assertContainsBits(final Collection<? extends Bit> bitList, final Bit... bits) {
